fix: disable osr for jeandle (#178)
## Related issue(s):

issue #173

## What this PR does / why we need it:

In the test case in #173, three threads are created to invoke
`addCounter`. But only the first thread enter the jeandle compiled
`addCounter`, because when the other two thread invoking `addCounter`,
there is a compilation task of `addCounter` on compile task queue, then
they just fall back to interpreter. When the loop in `addCounter`
trigger an OSR, the interpreter threads use Jeandle to compile an OSR
version of `addCounter`. Then Jeandle produce bad code because OSR is
unsupported now.

Signed-off-by: taofengliu <liutaofeng.ltf@antgroup.com>